+++ src/Application/Adapter/UuidGeneratorInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Proximum\Vimeet\Application\Adapter;
+
+interface UuidGeneratorInterface
+{
+    public function generate(): string;
+}