+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.42.0] - 2020-07-31
 ### Added
 - VIMEET-2168 - Video object on sheet template
 - Add fly system and google cloud storage adapter